% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectMain.R
\name{selectGenes}
\alias{selectGenes}
\title{Function to select genes from time course Abx RNAseq data}
\usage{
selectGenes(count_file, outdir, stag, l_cand = 1, base_lim = 50,
  use_t1 = FALSE, use_res = FALSE)
}
\arguments{
\item{count_file}{File containing read counts for genes for all samples}

\item{outdir}{Output directory for results and logs}

\item{stag}{A string appeded to the generated files}

\item{l_cand}{log2fold change limit for candidate genes [default : 1]}

\item{base_lim}{baseMean limits for a gene after DESeq2 [default: 50]}

\item{use_t1}{Set it to use first time point [default: FALSE]}

\item{use_res}{Set to use resistant strains [default: FALSE]}
}
\description{
This is the main function to generate genes from Abx RNAseq time course data
}
\examples{
selectGenes()
}
