% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSelectMtbMain.R
\name{selectGenesMtb}
\alias{selectGenesMtb}
\title{Function to select genes from Mtb time course Abx RNAseq data}
\usage{
selectGenesMtb(count_file, outdir, stag, sus, treated, untreated,
  timepoints, l_cand = 1, base_lim = 50, use_t1 = FALSE)
}
\arguments{
\item{count_file}{File containing read counts for genes for all samples}

\item{outdir}{Output directory for results and logs}

\item{stag}{A string appeded to the generated files}

\item{sus}{Susceptible strain}

\item{treated}{Condition/compound for treated samples}

\item{untreated}{Condition/control for untreated samples}

\item{timepoints}{timepoints in the time series data}

\item{l_cand}{log2fold change limit for candidate genes [default : 1]}

\item{base_lim}{baseMean limits for a gene after DESeq2 [default: 50]}

\item{use_t1}{Set it to use first time point [default: FALSE]}
}
\description{
This is the main function to generate genes from Mtb Abx RNAseq time course 
data
}
\examples{
selectGenesMtb()
}
